refactor: comment 관련 type 분리
